// src/glycemic_level/enum/glycemic_level.enum.ts
export enum TYPE_LEVEL {
    FAST = 'FAST', // GLICEMIA DE JEJUM
    RANDOM = 'RANDOM' // GLICEMIA ALEATÓRIA, AO LONGO DO DIA
}

export enum RATE {
    NORMAL = 'NORMAL',
    ALERT = 'ALERT',
    CRITICAL = 'CRITICAL'
}

export enum COLOR_RATE {
    GREEN = 'GREEN',
    YELLOW = 'YELLOW',
    RED = 'RED'
}